[AIRFLOW-5199] Move GCP Spanner to core
This commit moves GCP Spanner from contrib to core.
For more information check AIP-21.
