+++ tests/gcp/hooks/test_spanner.py
@@ -20,7 +20,7 @@
 import unittest
 
 from airflow import AirflowException
-from airflow.contrib.hooks.gcp_spanner_hook import CloudSpannerHook
+from airflow.gcp.hooks.spanner import CloudSpannerHook
 from tests.contrib.utils.base_gcp_mock import mock_base_gcp_hook_no_default_project_id, \
     GCP_PROJECT_ID_HOOK_UNIT_TEST, mock_base_gcp_hook_default_project_id
 from tests.compat import mock
@@ -37,7 +37,7 @@ class TestGcpSpannerHookDefaultProjectId(unittest.TestCase):
                         new=mock_base_gcp_hook_default_project_id):
             self.spanner_hook_default_project_id = CloudSpannerHook(gcp_conn_id='test')
 
-    @mock.patch('airflow.contrib.hooks.gcp_spanner_hook.CloudSpannerHook._get_client')
+    @mock.patch('airflow.gcp.hooks.spanner.CloudSpannerHook._get_client')
     def test_get_existing_instance(self, get_client):
         instance_method = get_client.return_value.instance
         instance_exists_method = instance_method.return_value.exists
@@ -48,7 +48,7 @@ class TestGcpSpannerHookDefaultProjectId(unittest.TestCase):
         instance_method.assert_called_once_with(instance_id='instance')
         self.assertIsNotNone(res)
 
-    @mock.patch('airflow.contrib.hooks.gcp_spanner_hook.CloudSpannerHook._get_client')
+    @mock.patch('airflow.gcp.hooks.spanner.CloudSpannerHook._get_client')
     def test_get_existing_instance_overridden_project_id(self, get_client):
         instance_method = get_client.return_value.instance
         instance_exists_method = instance_method.return_value.exists
@@ -59,7 +59,7 @@ class TestGcpSpannerHookDefaultProjectId(unittest.TestCase):
         instance_method.assert_called_once_with(instance_id='instance')
         self.assertIsNotNone(res)
 
-    @mock.patch('airflow.contrib.hooks.gcp_spanner_hook.CloudSpannerHook._get_client')
+    @mock.patch('airflow.gcp.hooks.spanner.CloudSpannerHook._get_client')
     def test_create_instance(self, get_client):
         instance_method = get_client.return_value.instance
         create_method = instance_method.return_value.create
@@ -77,7 +77,7 @@ class TestGcpSpannerHookDefaultProjectId(unittest.TestCase):
             node_count=1)
         self.assertIsNone(res)
 
-    @mock.patch('airflow.contrib.hooks.gcp_spanner_hook.CloudSpannerHook._get_client')
+    @mock.patch('airflow.gcp.hooks.spanner.CloudSpannerHook._get_client')
     def test_create_instance_overridden_project_id(self, get_client):
         instance_method = get_client.return_value.instance
         create_method = instance_method.return_value.create
@@ -96,7 +96,7 @@ class TestGcpSpannerHookDefaultProjectId(unittest.TestCase):
             node_count=1)
         self.assertIsNone(res)
 
-    @mock.patch('airflow.contrib.hooks.gcp_spanner_hook.CloudSpannerHook._get_client')
+    @mock.patch('airflow.gcp.hooks.spanner.CloudSpannerHook._get_client')
     def test_update_instance(self, get_client):
         instance_method = get_client.return_value.instance
         instance_exists_method = instance_method.return_value.exists
@@ -115,7 +115,7 @@ class TestGcpSpannerHookDefaultProjectId(unittest.TestCase):
         update_method.assert_called_with()
         self.assertIsNone(res)
 
-    @mock.patch('airflow.contrib.hooks.gcp_spanner_hook.CloudSpannerHook._get_client')
+    @mock.patch('airflow.gcp.hooks.spanner.CloudSpannerHook._get_client')
     def test_update_instance_overridden_project_id(self, get_client):
         instance_method = get_client.return_value.instance
         instance_exists_method = instance_method.return_value.exists
@@ -135,7 +135,7 @@ class TestGcpSpannerHookDefaultProjectId(unittest.TestCase):
         update_method.assert_called_with()
         self.assertIsNone(res)
 
-    @mock.patch('airflow.contrib.hooks.gcp_spanner_hook.CloudSpannerHook._get_client')
+    @mock.patch('airflow.gcp.hooks.spanner.CloudSpannerHook._get_client')
     def test_delete_instance(self, get_client):
         instance_method = get_client.return_value.instance
         instance_exists_method = instance_method.return_value.exists
@@ -150,7 +150,7 @@ class TestGcpSpannerHookDefaultProjectId(unittest.TestCase):
         delete_method.assert_called_with()
         self.assertIsNone(res)
 
-    @mock.patch('airflow.contrib.hooks.gcp_spanner_hook.CloudSpannerHook._get_client')
+    @mock.patch('airflow.gcp.hooks.spanner.CloudSpannerHook._get_client')
     def test_delete_instance_overridden_project_id(self, get_client):
         instance_method = get_client.return_value.instance
         instance_exists_method = instance_method.return_value.exists
@@ -166,7 +166,7 @@ class TestGcpSpannerHookDefaultProjectId(unittest.TestCase):
         delete_method.assert_called_with()
         self.assertIsNone(res)
 
-    @mock.patch('airflow.contrib.hooks.gcp_spanner_hook.CloudSpannerHook._get_client')
+    @mock.patch('airflow.gcp.hooks.spanner.CloudSpannerHook._get_client')
     def test_get_database(self, get_client):
         instance_method = get_client.return_value.instance
         instance_exists_method = instance_method.return_value.exists
@@ -183,7 +183,7 @@ class TestGcpSpannerHookDefaultProjectId(unittest.TestCase):
         database_exists_method.assert_called_with()
         self.assertIsNotNone(res)
 
-    @mock.patch('airflow.contrib.hooks.gcp_spanner_hook.CloudSpannerHook._get_client')
+    @mock.patch('airflow.gcp.hooks.spanner.CloudSpannerHook._get_client')
     def test_get_database_overridden_project_id(self, get_client):
         instance_method = get_client.return_value.instance
         instance_exists_method = instance_method.return_value.exists
@@ -201,7 +201,7 @@ class TestGcpSpannerHookDefaultProjectId(unittest.TestCase):
         database_exists_method.assert_called_with()
         self.assertIsNotNone(res)
 
-    @mock.patch('airflow.contrib.hooks.gcp_spanner_hook.CloudSpannerHook._get_client')
+    @mock.patch('airflow.gcp.hooks.spanner.CloudSpannerHook._get_client')
     def test_create_database(self, get_client):
         instance_method = get_client.return_value.instance
         instance_exists_method = instance_method.return_value.exists
@@ -218,7 +218,7 @@ class TestGcpSpannerHookDefaultProjectId(unittest.TestCase):
         database_create_method.assert_called_with()
         self.assertIsNone(res)
 
-    @mock.patch('airflow.contrib.hooks.gcp_spanner_hook.CloudSpannerHook._get_client')
+    @mock.patch('airflow.gcp.hooks.spanner.CloudSpannerHook._get_client')
     def test_create_database_overridden_project_id(self, get_client):
         instance_method = get_client.return_value.instance
         instance_exists_method = instance_method.return_value.exists
@@ -236,7 +236,7 @@ class TestGcpSpannerHookDefaultProjectId(unittest.TestCase):
         database_create_method.assert_called_with()
         self.assertIsNone(res)
 
-    @mock.patch('airflow.contrib.hooks.gcp_spanner_hook.CloudSpannerHook._get_client')
+    @mock.patch('airflow.gcp.hooks.spanner.CloudSpannerHook._get_client')
     def test_update_database(self, get_client):
         instance_method = get_client.return_value.instance
         instance_exists_method = instance_method.return_value.exists
@@ -253,7 +253,7 @@ class TestGcpSpannerHookDefaultProjectId(unittest.TestCase):
         database_update_ddl_method.assert_called_with(ddl_statements=[], operation_id=None)
         self.assertIsNone(res)
 
-    @mock.patch('airflow.contrib.hooks.gcp_spanner_hook.CloudSpannerHook._get_client')
+    @mock.patch('airflow.gcp.hooks.spanner.CloudSpannerHook._get_client')
     def test_update_database_overridden_project_id(self, get_client):
         instance_method = get_client.return_value.instance
         instance_exists_method = instance_method.return_value.exists
@@ -271,7 +271,7 @@ class TestGcpSpannerHookDefaultProjectId(unittest.TestCase):
         database_update_ddl_method.assert_called_with(ddl_statements=[], operation_id=None)
         self.assertIsNone(res)
 
-    @mock.patch('airflow.contrib.hooks.gcp_spanner_hook.CloudSpannerHook._get_client')
+    @mock.patch('airflow.gcp.hooks.spanner.CloudSpannerHook._get_client')
     def test_delete_database(self, get_client):
         instance_method = get_client.return_value.instance
         instance_exists_method = instance_method.return_value.exists
@@ -290,7 +290,7 @@ class TestGcpSpannerHookDefaultProjectId(unittest.TestCase):
         database_drop_method.assert_called_with()
         self.assertTrue(res)
 
-    @mock.patch('airflow.contrib.hooks.gcp_spanner_hook.CloudSpannerHook._get_client')
+    @mock.patch('airflow.gcp.hooks.spanner.CloudSpannerHook._get_client')
     def test_delete_database_overridden_project_id(self, get_client):
         instance_method = get_client.return_value.instance
         instance_exists_method = instance_method.return_value.exists
@@ -310,7 +310,7 @@ class TestGcpSpannerHookDefaultProjectId(unittest.TestCase):
         database_drop_method.assert_called_with()
         self.assertTrue(res)
 
-    @mock.patch('airflow.contrib.hooks.gcp_spanner_hook.CloudSpannerHook._get_client')
+    @mock.patch('airflow.gcp.hooks.spanner.CloudSpannerHook._get_client')
     def test_execute_dml(self, get_client):
         instance_method = get_client.return_value.instance
         instance_exists_method = instance_method.return_value.exists
@@ -327,7 +327,7 @@ class TestGcpSpannerHookDefaultProjectId(unittest.TestCase):
         run_in_transaction_method.assert_called_with(mock.ANY)
         self.assertIsNone(res)
 
-    @mock.patch('airflow.contrib.hooks.gcp_spanner_hook.CloudSpannerHook._get_client')
+    @mock.patch('airflow.gcp.hooks.spanner.CloudSpannerHook._get_client')
     def test_execute_dml_overridden_project_id(self, get_client):
         instance_method = get_client.return_value.instance
         instance_exists_method = instance_method.return_value.exists
@@ -353,7 +353,7 @@ class TestGcpSpannerHookNoDefaultProjectID(unittest.TestCase):
                         new=mock_base_gcp_hook_no_default_project_id):
             self.spanner_hook_no_default_project_id = CloudSpannerHook(gcp_conn_id='test')
 
-    @mock.patch('airflow.contrib.hooks.gcp_spanner_hook.CloudSpannerHook._get_client')
+    @mock.patch('airflow.gcp.hooks.spanner.CloudSpannerHook._get_client')
     def test_get_existing_instance_missing_project_id(self, get_client):
         instance_method = get_client.return_value.instance
         instance_exists_method = instance_method.return_value.exists
@@ -366,7 +366,7 @@ class TestGcpSpannerHookNoDefaultProjectID(unittest.TestCase):
         err = cm.exception
         self.assertIn("The project id must be passed", str(err))
 
-    @mock.patch('airflow.contrib.hooks.gcp_spanner_hook.CloudSpannerHook._get_client')
+    @mock.patch('airflow.gcp.hooks.spanner.CloudSpannerHook._get_client')
     def test_get_existing_instance_overridden_project_id(self, get_client):
         instance_method = get_client.return_value.instance
         instance_exists_method = instance_method.return_value.exists
@@ -377,7 +377,7 @@ class TestGcpSpannerHookNoDefaultProjectID(unittest.TestCase):
         instance_method.assert_called_once_with(instance_id='instance')
         self.assertIsNotNone(res)
 
-    @mock.patch('airflow.contrib.hooks.gcp_spanner_hook.CloudSpannerHook._get_client')
+    @mock.patch('airflow.gcp.hooks.spanner.CloudSpannerHook._get_client')
     def test_get_non_existing_instance(self, get_client):
         instance_method = get_client.return_value.instance
         instance_exists_method = instance_method.return_value.exists
@@ -388,7 +388,7 @@ class TestGcpSpannerHookNoDefaultProjectID(unittest.TestCase):
         instance_method.assert_called_once_with(instance_id='instance')
         self.assertIsNone(res)
 
-    @mock.patch('airflow.contrib.hooks.gcp_spanner_hook.CloudSpannerHook._get_client')
+    @mock.patch('airflow.gcp.hooks.spanner.CloudSpannerHook._get_client')
     def test_create_instance_missing_project_id(self, get_client):
         instance_method = get_client.return_value.instance
         create_method = instance_method.return_value.create
@@ -406,7 +406,7 @@ class TestGcpSpannerHookNoDefaultProjectID(unittest.TestCase):
         err = cm.exception
         self.assertIn("The project id must be passed", str(err))
 
-    @mock.patch('airflow.contrib.hooks.gcp_spanner_hook.CloudSpannerHook._get_client')
+    @mock.patch('airflow.gcp.hooks.spanner.CloudSpannerHook._get_client')
     def test_create_instance_overridden_project_id(self, get_client):
         instance_method = get_client.return_value.instance
         create_method = instance_method.return_value.create
@@ -425,7 +425,7 @@ class TestGcpSpannerHookNoDefaultProjectID(unittest.TestCase):
             node_count=1)
         self.assertIsNone(res)
 
-    @mock.patch('airflow.contrib.hooks.gcp_spanner_hook.CloudSpannerHook._get_client')
+    @mock.patch('airflow.gcp.hooks.spanner.CloudSpannerHook._get_client')
     def test_update_instance_missing_project_id(self, get_client):
         instance_method = get_client.return_value.instance
         instance_exists_method = instance_method.return_value.exists
@@ -444,7 +444,7 @@ class TestGcpSpannerHookNoDefaultProjectID(unittest.TestCase):
         err = cm.exception
         self.assertIn("The project id must be passed", str(err))
 
-    @mock.patch('airflow.contrib.hooks.gcp_spanner_hook.CloudSpannerHook._get_client')
+    @mock.patch('airflow.gcp.hooks.spanner.CloudSpannerHook._get_client')
     def test_update_instance_overridden_project_id(self, get_client):
         instance_method = get_client.return_value.instance
         instance_exists_method = instance_method.return_value.exists
@@ -464,7 +464,7 @@ class TestGcpSpannerHookNoDefaultProjectID(unittest.TestCase):
         update_method.assert_called_with()
         self.assertIsNone(res)
 
-    @mock.patch('airflow.contrib.hooks.gcp_spanner_hook.CloudSpannerHook._get_client')
+    @mock.patch('airflow.gcp.hooks.spanner.CloudSpannerHook._get_client')
     def test_delete_instance_missing_project_id(self, get_client):
         instance_method = get_client.return_value.instance
         instance_exists_method = instance_method.return_value.exists
@@ -480,7 +480,7 @@ class TestGcpSpannerHookNoDefaultProjectID(unittest.TestCase):
         err = cm.exception
         self.assertIn("The project id must be passed", str(err))
 
-    @mock.patch('airflow.contrib.hooks.gcp_spanner_hook.CloudSpannerHook._get_client')
+    @mock.patch('airflow.gcp.hooks.spanner.CloudSpannerHook._get_client')
     def test_delete_instance_overridden_project_id(self, get_client):
         instance_method = get_client.return_value.instance
         instance_exists_method = instance_method.return_value.exists
@@ -496,7 +496,7 @@ class TestGcpSpannerHookNoDefaultProjectID(unittest.TestCase):
         delete_method.assert_called_with()
         self.assertIsNone(res)
 
-    @mock.patch('airflow.contrib.hooks.gcp_spanner_hook.CloudSpannerHook._get_client')
+    @mock.patch('airflow.gcp.hooks.spanner.CloudSpannerHook._get_client')
     def test_get_database_missing_project_id(self, get_client):
         instance_method = get_client.return_value.instance
         instance_exists_method = instance_method.return_value.exists
@@ -514,7 +514,7 @@ class TestGcpSpannerHookNoDefaultProjectID(unittest.TestCase):
         err = cm.exception
         self.assertIn("The project id must be passed", str(err))
 
-    @mock.patch('airflow.contrib.hooks.gcp_spanner_hook.CloudSpannerHook._get_client')
+    @mock.patch('airflow.gcp.hooks.spanner.CloudSpannerHook._get_client')
     def test_get_database_overridden_project_id(self, get_client):
         instance_method = get_client.return_value.instance
         instance_exists_method = instance_method.return_value.exists
@@ -532,7 +532,7 @@ class TestGcpSpannerHookNoDefaultProjectID(unittest.TestCase):
         database_exists_method.assert_called_with()
         self.assertIsNotNone(res)
 
-    @mock.patch('airflow.contrib.hooks.gcp_spanner_hook.CloudSpannerHook._get_client')
+    @mock.patch('airflow.gcp.hooks.spanner.CloudSpannerHook._get_client')
     def test_create_database_missing_project_id(self, get_client):
         instance_method = get_client.return_value.instance
         instance_exists_method = instance_method.return_value.exists
@@ -551,7 +551,7 @@ class TestGcpSpannerHookNoDefaultProjectID(unittest.TestCase):
         err = cm.exception
         self.assertIn("The project id must be passed", str(err))
 
-    @mock.patch('airflow.contrib.hooks.gcp_spanner_hook.CloudSpannerHook._get_client')
+    @mock.patch('airflow.gcp.hooks.spanner.CloudSpannerHook._get_client')
     def test_create_database_overridden_project_id(self, get_client):
         instance_method = get_client.return_value.instance
         instance_exists_method = instance_method.return_value.exists
@@ -569,7 +569,7 @@ class TestGcpSpannerHookNoDefaultProjectID(unittest.TestCase):
         database_create_method.assert_called_with()
         self.assertIsNone(res)
 
-    @mock.patch('airflow.contrib.hooks.gcp_spanner_hook.CloudSpannerHook._get_client')
+    @mock.patch('airflow.gcp.hooks.spanner.CloudSpannerHook._get_client')
     def test_update_database_missing_project_id(self, get_client):
         instance_method = get_client.return_value.instance
         instance_exists_method = instance_method.return_value.exists
@@ -588,7 +588,7 @@ class TestGcpSpannerHookNoDefaultProjectID(unittest.TestCase):
         err = cm.exception
         self.assertIn("The project id must be passed", str(err))
 
-    @mock.patch('airflow.contrib.hooks.gcp_spanner_hook.CloudSpannerHook._get_client')
+    @mock.patch('airflow.gcp.hooks.spanner.CloudSpannerHook._get_client')
     def test_update_database_overridden_project_id(self, get_client):
         instance_method = get_client.return_value.instance
         instance_exists_method = instance_method.return_value.exists
@@ -606,7 +606,7 @@ class TestGcpSpannerHookNoDefaultProjectID(unittest.TestCase):
         database_update_ddl_method.assert_called_with(ddl_statements=[], operation_id=None)
         self.assertIsNone(res)
 
-    @mock.patch('airflow.contrib.hooks.gcp_spanner_hook.CloudSpannerHook._get_client')
+    @mock.patch('airflow.gcp.hooks.spanner.CloudSpannerHook._get_client')
     def test_update_database_overridden_project_id_and_operation(self, get_client):
         instance_method = get_client.return_value.instance
         instance_exists_method = instance_method.return_value.exists
@@ -625,7 +625,7 @@ class TestGcpSpannerHookNoDefaultProjectID(unittest.TestCase):
         database_update_ddl_method.assert_called_with(ddl_statements=[], operation_id="operation")
         self.assertIsNone(res)
 
-    @mock.patch('airflow.contrib.hooks.gcp_spanner_hook.CloudSpannerHook._get_client')
+    @mock.patch('airflow.gcp.hooks.spanner.CloudSpannerHook._get_client')
     def test_delete_database_missing_project_id(self, get_client):
         instance_method = get_client.return_value.instance
         instance_exists_method = instance_method.return_value.exists
@@ -646,7 +646,7 @@ class TestGcpSpannerHookNoDefaultProjectID(unittest.TestCase):
         err = cm.exception
         self.assertIn("The project id must be passed", str(err))
 
-    @mock.patch('airflow.contrib.hooks.gcp_spanner_hook.CloudSpannerHook._get_client')
+    @mock.patch('airflow.gcp.hooks.spanner.CloudSpannerHook._get_client')
     def test_delete_database_overridden_project_id(self, get_client):
         instance_method = get_client.return_value.instance
         instance_exists_method = instance_method.return_value.exists
@@ -666,7 +666,7 @@ class TestGcpSpannerHookNoDefaultProjectID(unittest.TestCase):
         database_drop_method.assert_called_with()
         self.assertTrue(res)
 
-    @mock.patch('airflow.contrib.hooks.gcp_spanner_hook.CloudSpannerHook._get_client')
+    @mock.patch('airflow.gcp.hooks.spanner.CloudSpannerHook._get_client')
     def test_delete_database_missing_database(self, get_client):
         instance_method = get_client.return_value.instance
         instance_exists_method = instance_method.return_value.exists
@@ -685,7 +685,7 @@ class TestGcpSpannerHookNoDefaultProjectID(unittest.TestCase):
         database_exists_method.assert_called_with()
         database_drop_method.assert_not_called()
 
-    @mock.patch('airflow.contrib.hooks.gcp_spanner_hook.CloudSpannerHook._get_client')
+    @mock.patch('airflow.gcp.hooks.spanner.CloudSpannerHook._get_client')
     def test_execute_dml_missing_project_id(self, get_client):
         instance_method = get_client.return_value.instance
         instance_exists_method = instance_method.return_value.exists
@@ -704,7 +704,7 @@ class TestGcpSpannerHookNoDefaultProjectID(unittest.TestCase):
         err = cm.exception
         self.assertIn("The project id must be passed", str(err))
 
-    @mock.patch('airflow.contrib.hooks.gcp_spanner_hook.CloudSpannerHook._get_client')
+    @mock.patch('airflow.gcp.hooks.spanner.CloudSpannerHook._get_client')
     def test_execute_dml_overridden_project_id(self, get_client):
         instance_method = get_client.return_value.instance
         instance_exists_method = instance_method.return_value.exists
