+++ tests/gcp/operators/test_spanner.py
@@ -21,7 +21,7 @@ import unittest
 from parameterized import parameterized
 
 from airflow import AirflowException
-from airflow.contrib.operators.gcp_spanner_operator import \
+from airflow.gcp.operators.spanner import \
     CloudSpannerInstanceDeployOperator, \
     CloudSpannerInstanceDeleteOperator, \
     CloudSpannerInstanceDatabaseQueryOperator, \
@@ -45,7 +45,7 @@ DDL_STATEMENTS = [CREATE_QUERY, CREATE_QUERY_2]
 
 
 class CloudSpannerTest(unittest.TestCase):
-    @mock.patch("airflow.contrib.operators.gcp_spanner_operator.CloudSpannerHook")
+    @mock.patch("airflow.gcp.operators.spanner.CloudSpannerHook")
     def test_instance_create(self, mock_hook):
         mock_hook.return_value.get_instance.return_value = None
         op = CloudSpannerInstanceDeployOperator(
@@ -68,7 +68,7 @@ class CloudSpannerTest(unittest.TestCase):
         mock_hook.return_value.update_instance.assert_not_called()
         self.assertIsNone(result)
 
-    @mock.patch("airflow.contrib.operators.gcp_spanner_operator.CloudSpannerHook")
+    @mock.patch("airflow.gcp.operators.spanner.CloudSpannerHook")
     def test_instance_create_missing_project_id(self, mock_hook):
         mock_hook.return_value.get_instance.return_value = None
         op = CloudSpannerInstanceDeployOperator(
@@ -90,7 +90,7 @@ class CloudSpannerTest(unittest.TestCase):
         mock_hook.return_value.update_instance.assert_not_called()
         self.assertIsNone(result)
 
-    @mock.patch("airflow.contrib.operators.gcp_spanner_operator.CloudSpannerHook")
+    @mock.patch("airflow.gcp.operators.spanner.CloudSpannerHook")
     def test_instance_update(self, mock_hook):
         mock_hook.return_value.get_instance.return_value = {"name": INSTANCE_ID}
         op = CloudSpannerInstanceDeployOperator(
@@ -113,7 +113,7 @@ class CloudSpannerTest(unittest.TestCase):
         mock_hook.return_value.create_instance.assert_not_called()
         self.assertIsNone(result)
 
-    @mock.patch("airflow.contrib.operators.gcp_spanner_operator.CloudSpannerHook")
+    @mock.patch("airflow.gcp.operators.spanner.CloudSpannerHook")
     def test_instance_update_missing_project_id(self, mock_hook):
         mock_hook.return_value.get_instance.return_value = {"name": INSTANCE_ID}
         op = CloudSpannerInstanceDeployOperator(
@@ -135,7 +135,7 @@ class CloudSpannerTest(unittest.TestCase):
         mock_hook.return_value.create_instance.assert_not_called()
         self.assertIsNone(result)
 
-    @mock.patch("airflow.contrib.operators.gcp_spanner_operator.CloudSpannerHook")
+    @mock.patch("airflow.gcp.operators.spanner.CloudSpannerHook")
     def test_instance_create_aborts_and_succeeds_if_instance_exists(self, mock_hook):
         mock_hook.return_value.get_instance.return_value = {"name": INSTANCE_ID}
         op = CloudSpannerInstanceDeployOperator(
@@ -155,7 +155,7 @@ class CloudSpannerTest(unittest.TestCase):
         ("", INSTANCE_ID, "project_id"),
         (PROJECT_ID, "", "instance_id"),
     ])
-    @mock.patch("airflow.contrib.operators.gcp_spanner_operator.CloudSpannerHook")
+    @mock.patch("airflow.gcp.operators.spanner.CloudSpannerHook")
     def test_instance_create_ex_if_param_missing(self, project_id, instance_id,
                                                  exp_msg, mock_hook):
         with self.assertRaises(AirflowException) as cm:
@@ -171,7 +171,7 @@ class CloudSpannerTest(unittest.TestCase):
         self.assertIn("The required parameter '{}' is empty".format(exp_msg), str(err))
         mock_hook.assert_not_called()
 
-    @mock.patch("airflow.contrib.operators.gcp_spanner_operator.CloudSpannerHook")
+    @mock.patch("airflow.gcp.operators.spanner.CloudSpannerHook")
     def test_instance_delete(self, mock_hook):
         mock_hook.return_value.get_instance.return_value = {"name": INSTANCE_ID}
         op = CloudSpannerInstanceDeleteOperator(
@@ -186,7 +186,7 @@ class CloudSpannerTest(unittest.TestCase):
         )
         self.assertTrue(result)
 
-    @mock.patch("airflow.contrib.operators.gcp_spanner_operator.CloudSpannerHook")
+    @mock.patch("airflow.gcp.operators.spanner.CloudSpannerHook")
     def test_instance_delete_missing_project_id(self, mock_hook):
         mock_hook.return_value.get_instance.return_value = {"name": INSTANCE_ID}
         op = CloudSpannerInstanceDeleteOperator(
@@ -201,7 +201,7 @@ class CloudSpannerTest(unittest.TestCase):
         )
         self.assertTrue(result)
 
-    @mock.patch("airflow.contrib.operators.gcp_spanner_operator.CloudSpannerHook")
+    @mock.patch("airflow.gcp.operators.spanner.CloudSpannerHook")
     def test_instance_delete_aborts_and_succeeds_if_instance_does_not_exist(self,
                                                                             mock_hook):
         mock_hook.return_value.get_instance.return_value = None
@@ -219,7 +219,7 @@ class CloudSpannerTest(unittest.TestCase):
         ("", INSTANCE_ID, "project_id"),
         (PROJECT_ID, "", "instance_id"),
     ])
-    @mock.patch("airflow.contrib.operators.gcp_spanner_operator.CloudSpannerHook")
+    @mock.patch("airflow.gcp.operators.spanner.CloudSpannerHook")
     def test_instance_delete_ex_if_param_missing(self, project_id, instance_id, exp_msg,
                                                  mock_hook):
         with self.assertRaises(AirflowException) as cm:
@@ -232,7 +232,7 @@ class CloudSpannerTest(unittest.TestCase):
         self.assertIn("The required parameter '{}' is empty".format(exp_msg), str(err))
         mock_hook.assert_not_called()
 
-    @mock.patch("airflow.contrib.operators.gcp_spanner_operator.CloudSpannerHook")
+    @mock.patch("airflow.gcp.operators.spanner.CloudSpannerHook")
     def test_instance_query(self, mock_hook):
         mock_hook.return_value.execute_sql.return_value = None
         op = CloudSpannerInstanceDatabaseQueryOperator(
@@ -251,7 +251,7 @@ class CloudSpannerTest(unittest.TestCase):
         )
         self.assertIsNone(result)
 
-    @mock.patch("airflow.contrib.operators.gcp_spanner_operator.CloudSpannerHook")
+    @mock.patch("airflow.gcp.operators.spanner.CloudSpannerHook")
     def test_instance_query_missing_project_id(self, mock_hook):
         mock_hook.return_value.execute_sql.return_value = None
         op = CloudSpannerInstanceDatabaseQueryOperator(
@@ -274,7 +274,7 @@ class CloudSpannerTest(unittest.TestCase):
         (PROJECT_ID, INSTANCE_ID, "", INSERT_QUERY, "database_id"),
         (PROJECT_ID, INSTANCE_ID, DB_ID, "", "query"),
     ])
-    @mock.patch("airflow.contrib.operators.gcp_spanner_operator.CloudSpannerHook")
+    @mock.patch("airflow.gcp.operators.spanner.CloudSpannerHook")
     def test_instance_query_ex_if_param_missing(self, project_id, instance_id,
                                                 database_id, query, exp_msg, mock_hook):
         with self.assertRaises(AirflowException) as cm:
@@ -289,7 +289,7 @@ class CloudSpannerTest(unittest.TestCase):
         self.assertIn("The required parameter '{}' is empty".format(exp_msg), str(err))
         mock_hook.assert_not_called()
 
-    @mock.patch("airflow.contrib.operators.gcp_spanner_operator.CloudSpannerHook")
+    @mock.patch("airflow.gcp.operators.spanner.CloudSpannerHook")
     def test_instance_query_dml(self, mock_hook):
         mock_hook.return_value.execute_dml.return_value = None
         op = CloudSpannerInstanceDatabaseQueryOperator(
@@ -305,7 +305,7 @@ class CloudSpannerTest(unittest.TestCase):
             project_id=PROJECT_ID, instance_id=INSTANCE_ID, database_id=DB_ID, queries=[INSERT_QUERY]
         )
 
-    @mock.patch("airflow.contrib.operators.gcp_spanner_operator.CloudSpannerHook")
+    @mock.patch("airflow.gcp.operators.spanner.CloudSpannerHook")
     def test_instance_query_dml_list(self, mock_hook):
         mock_hook.return_value.execute_dml.return_value = None
         op = CloudSpannerInstanceDatabaseQueryOperator(
@@ -322,7 +322,7 @@ class CloudSpannerTest(unittest.TestCase):
             database_id=DB_ID, queries=[INSERT_QUERY, INSERT_QUERY_2]
         )
 
-    @mock.patch("airflow.contrib.operators.gcp_spanner_operator.CloudSpannerHook")
+    @mock.patch("airflow.gcp.operators.spanner.CloudSpannerHook")
     def test_database_create(self, mock_hook):
         mock_hook.return_value.get_database.return_value = None
         op = CloudSpannerInstanceDatabaseDeployOperator(
@@ -341,7 +341,7 @@ class CloudSpannerTest(unittest.TestCase):
         mock_hook.return_value.update_database.assert_not_called()
         self.assertTrue(result)
 
-    @mock.patch("airflow.contrib.operators.gcp_spanner_operator.CloudSpannerHook")
+    @mock.patch("airflow.gcp.operators.spanner.CloudSpannerHook")
     def test_database_create_missing_project_id(self, mock_hook):
         mock_hook.return_value.get_database.return_value = None
         op = CloudSpannerInstanceDatabaseDeployOperator(
@@ -359,7 +359,7 @@ class CloudSpannerTest(unittest.TestCase):
         mock_hook.return_value.update_database.assert_not_called()
         self.assertTrue(result)
 
-    @mock.patch("airflow.contrib.operators.gcp_spanner_operator.CloudSpannerHook")
+    @mock.patch("airflow.gcp.operators.spanner.CloudSpannerHook")
     def test_database_create_with_pre_existing_db(self, mock_hook):
         mock_hook.return_value.get_database.return_value = {"name": DB_ID}
         op = CloudSpannerInstanceDatabaseDeployOperator(
@@ -380,7 +380,7 @@ class CloudSpannerTest(unittest.TestCase):
         (PROJECT_ID, "", DB_ID, DDL_STATEMENTS, 'instance_id'),
         (PROJECT_ID, INSTANCE_ID, "", DDL_STATEMENTS, 'database_id'),
     ])
-    @mock.patch("airflow.contrib.operators.gcp_spanner_operator.CloudSpannerHook")
+    @mock.patch("airflow.gcp.operators.spanner.CloudSpannerHook")
     def test_database_create_ex_if_param_missing(self,
                                                  project_id, instance_id,
                                                  database_id, ddl_statements,
@@ -397,7 +397,7 @@ class CloudSpannerTest(unittest.TestCase):
         self.assertIn("The required parameter '{}' is empty".format(exp_msg), str(err))
         mock_hook.assert_not_called()
 
-    @mock.patch("airflow.contrib.operators.gcp_spanner_operator.CloudSpannerHook")
+    @mock.patch("airflow.gcp.operators.spanner.CloudSpannerHook")
     def test_database_update(self, mock_hook):
         mock_hook.return_value.get_database.return_value = {"name": DB_ID}
         op = CloudSpannerInstanceDatabaseUpdateOperator(
@@ -415,7 +415,7 @@ class CloudSpannerTest(unittest.TestCase):
         )
         self.assertTrue(result)
 
-    @mock.patch("airflow.contrib.operators.gcp_spanner_operator.CloudSpannerHook")
+    @mock.patch("airflow.gcp.operators.spanner.CloudSpannerHook")
     def test_database_update_missing_project_id(self, mock_hook):
         mock_hook.return_value.get_database.return_value = {"name": DB_ID}
         op = CloudSpannerInstanceDatabaseUpdateOperator(
@@ -437,7 +437,7 @@ class CloudSpannerTest(unittest.TestCase):
         (PROJECT_ID, "", DB_ID, DDL_STATEMENTS, 'instance_id'),
         (PROJECT_ID, INSTANCE_ID, "", DDL_STATEMENTS, 'database_id'),
     ])
-    @mock.patch("airflow.contrib.operators.gcp_spanner_operator.CloudSpannerHook")
+    @mock.patch("airflow.gcp.operators.spanner.CloudSpannerHook")
     def test_database_update_ex_if_param_missing(self, project_id, instance_id,
                                                  database_id, ddl_statements,
                                                  exp_msg, mock_hook):
@@ -453,7 +453,7 @@ class CloudSpannerTest(unittest.TestCase):
         self.assertIn("The required parameter '{}' is empty".format(exp_msg), str(err))
         mock_hook.assert_not_called()
 
-    @mock.patch("airflow.contrib.operators.gcp_spanner_operator.CloudSpannerHook")
+    @mock.patch("airflow.gcp.operators.spanner.CloudSpannerHook")
     def test_database_update_ex_if_database_not_exist(self, mock_hook):
         mock_hook.return_value.get_database.return_value = None
         with self.assertRaises(AirflowException) as cm:
@@ -470,7 +470,7 @@ class CloudSpannerTest(unittest.TestCase):
                       "instance 'instance-id' is missing", str(err))
         mock_hook.assert_called_once_with(gcp_conn_id="google_cloud_default")
 
-    @mock.patch("airflow.contrib.operators.gcp_spanner_operator.CloudSpannerHook")
+    @mock.patch("airflow.gcp.operators.spanner.CloudSpannerHook")
     def test_database_delete(self, mock_hook):
         mock_hook.return_value.get_database.return_value = {"name": DB_ID}
         op = CloudSpannerInstanceDatabaseDeleteOperator(
@@ -486,7 +486,7 @@ class CloudSpannerTest(unittest.TestCase):
         )
         self.assertTrue(result)
 
-    @mock.patch("airflow.contrib.operators.gcp_spanner_operator.CloudSpannerHook")
+    @mock.patch("airflow.gcp.operators.spanner.CloudSpannerHook")
     def test_database_delete_missing_project_id(self, mock_hook):
         mock_hook.return_value.get_database.return_value = {"name": DB_ID}
         op = CloudSpannerInstanceDatabaseDeleteOperator(
@@ -501,7 +501,7 @@ class CloudSpannerTest(unittest.TestCase):
         )
         self.assertTrue(result)
 
-    @mock.patch("airflow.contrib.operators.gcp_spanner_operator.CloudSpannerHook")
+    @mock.patch("airflow.gcp.operators.spanner.CloudSpannerHook")
     def test_database_delete_exits_and_succeeds_if_database_does_not_exist(self,
                                                                            mock_hook):
         mock_hook.return_value.get_database.return_value = None
@@ -521,7 +521,7 @@ class CloudSpannerTest(unittest.TestCase):
         (PROJECT_ID, "", DB_ID, DDL_STATEMENTS, 'instance_id'),
         (PROJECT_ID, INSTANCE_ID, "", DDL_STATEMENTS, 'database_id'),
     ])
-    @mock.patch("airflow.contrib.operators.gcp_spanner_operator.CloudSpannerHook")
+    @mock.patch("airflow.gcp.operators.spanner.CloudSpannerHook")
     def test_database_delete_ex_if_param_missing(self, project_id, instance_id,
                                                  database_id, ddl_statements,
                                                  exp_msg, mock_hook):
